Implemented single/multiple switch in maxwell, and changed executer.

+++ milonni.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python3
+#!/usr/bin/python
 
 import scipy.integrate
 import numpy as np
@@ -208,7 +208,7 @@ ax.annotate('B', xy=(instabPointsBN[0][0], instabPointsBN[1][0]), xytext=(instab
 
 plt.draw()
 
-anim = ani.FuncAnimation(fig, step(G, k, f, pMin, pSteps), init_func=init, frames=math.ceil((pMax-pMin)/pSteps), blit=False) #, interval=10, fargs=(q, status)
+anim = ani.FuncAnimation(fig, step(G, k, f, pMin, pSteps), init_func=init, frames=int(math.ceil((pMax-pMin)/pSteps)), blit=False) #, interval=10, fargs=(q, status)
 
 plt.show()
